Use context manager instead of finalizer to free resources
It is better to use the context manager to free resources, because
the `__del__()` method might not be called for objects that exists
when the interpreter exits. The `del` also does not directly call
the `__del__()` method, it just decreases the reference counter.

Another advantage is that one does not need to remember to free the
resources, they are freed automaticaly when object exits the `with`
block.

https://docs.python.org/3/reference/datamodel.html#object.__del__