Update certifi dependency
Loosen restrictions for certifi and urllib3

There is a certifi vulnerability warning which is fixed
by a newer version than the previously pinned one. Opening up the dependency
of certifi and urllib3 should also resolve
future issues when it coms to these packages.
Also loosening up types requests as this is handled
by the requests library too.